Move object conversion tests to their own directory
