+++ tests/test_conversion.py
@@ -1,25 +1,6 @@
-import voluptuous
-
 from voluptuary import to_voluptuous
 from .conversion import check_jsonschema_validation
 from .conversion import check_voluptuous_validation
-
-
-def test_empty_schema():
-    schema_in = {}
-    schema_out = to_voluptuous(schema_in)
-    for target in (
-        None,
-        '',
-        False,
-        42,
-        "I'm a string",
-        {},
-        [],
-        {"an": ["arbitrarily", "nested"], "data": "structure"},
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=True)
-        check_voluptuous_validation(schema_out, target, should_validate=True)
 
 
 def test_type_string_schema():
@@ -117,141 +98,6 @@ def test_list_of_types_schema():
         None,
         {},
         [],
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=False)
-        check_voluptuous_validation(schema_out, target, should_validate=False)
-
-
-def test_type_object_schema():
-    schema_in = {'type': 'object'}
-    schema_out = to_voluptuous(schema_in)
-
-    for target in (
-        {},
-        {'a': 1},
-        {'a': {'a': {'a': 1}}},
-        {'a': [{'a': {'a': 1}}]},
-        # Using non-string keys is invalid JSON. Technically, this shouldn't
-        # validate. However, we match the behavior of `jsonschema`.
-        {1: 1},
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=True)
-        check_voluptuous_validation(schema_out, target, should_validate=True)
-
-    for target in (
-        None,
-        [],
-        ['wumbo'],
-        [{"wumbo": 'abc'}],
-        1,
-        '',
-        1.5,
-        True,
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=False)
-        check_voluptuous_validation(schema_out, target, should_validate=False)
-
-
-def test_type_object_with_properties():
-    schema_in = {
-        'type': 'object', 'properties': {
-            'a-string': {'type': 'string'},
-            'an-integer': {'type': 'integer'},
-        }
-    }
-    schema_out = to_voluptuous(schema_in)
-
-    for target in (
-        {},
-        {'a-string': 'a'},
-        {'an-integer': 1},
-        {'a-string': '', 'an-integer': 0},
-        {'a-string': 'a', 'an-integer': 123},
-        {'extra-key': 1},
-        {'a-string': 'abc', 'extra-key': 1},
-        {'a-string': 'abc', 'an-integer': 1234, 'extra-key': 1},
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=True)
-        check_voluptuous_validation(schema_out, target, should_validate=True)
-
-    for target in (
-        {'a-string': {}},
-        {'a-string': 1},
-        {'a-string': 'a', 'an-integer': 'adsf'},
-        {'a-string': 'a', 'an-integer': 1.5},
-        None,
-        1,
-        1.5,
-        'abc',
-        False,
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=False)
-        check_voluptuous_validation(schema_out, target, should_validate=False)
-
-
-def test_type_object_with_additional_properties_false():
-    schema_in = {
-        'type': 'object',
-        'properties': {'a': {'type': 'null'}},
-        'additionalProperties': False,
-    }
-    schema_out = to_voluptuous(schema_in)
-
-    for target in (
-        {},
-        {'a': None},
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=True)
-        check_voluptuous_validation(schema_out, target, should_validate=True)
-
-    for target in (
-        {'b': None},
-        {'a': None, 'b': None},
-        None,
-        1,
-        1.5,
-        'abc',
-        False,
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=False)
-        check_voluptuous_validation(schema_out, target, should_validate=False)
-
-
-def test_type_object_with_additional_properties_schema():
-    schema_in = {
-        'type': 'object',
-        'properties': {'a': {'type': 'boolean'}},
-        'additionalProperties': {'type': 'string'},
-    }
-    schema_out = to_voluptuous(schema_in)
-    print(schema_out)
-
-    works = voluptuous.Schema(
-        {
-            'a': voluptuous.Schema(bool),
-            voluptuous.Extra: voluptuous.Schema(str),
-        }
-    )
-    print(works)
-
-    for target in (
-        {},
-        {'a': True},
-        {'a': False},
-        {'b': 'abc'},
-        {'a': True, 'b': 'abc'},
-        {'a': False, 'b': 'abc'},
-        {'a': False, 'b': 'abc', 1: 'adf'},
-    ):
-        check_jsonschema_validation(schema_in, target, should_validate=True)
-        check_voluptuous_validation(schema_out, target, should_validate=True)
-
-    for target in (
-        {'b': False},
-        {'b': 1},
-        {'b': 1.5},
-        {'b': {'asdf': 'asdf'}},
-        {'b': ['asdf']},
     ):
         check_jsonschema_validation(schema_in, target, should_validate=False)
         check_voluptuous_validation(schema_out, target, should_validate=False)
